// application/frontend/src/app/core/models/column.ts
/**
 * @license
 * Copyright 2022 Google LLC
 *
 * Use of this source code is governed by an MIT-style
 * license that can be found in the LICENSE file or at
 * https://opensource.org/licenses/MIT.
 */

export interface Column {
  id: string;
  label: string;
  active?: boolean;
  toggleable?: boolean;
  toggleableHidden?: boolean;
}